refactor(structure): update library structure
