+++ src/data-structures/sequences/linked-list-node.ts
