--- src/data-structures/sequences/queues/index.ts
@@ -1,2 +1,0 @@
-export { Queue } from './queue/index.js';
-export { Deque } from './deque/index.js';
