+++ src/data-structures/sequences/stack.ts
@@ -1,5 +1,5 @@
-import { LinkedList } from '../linked-list/index.js';
-import { type TypedEqualityFunction } from '../../utils.js';
+import { LinkedList } from './linked-list';
+import { type TypedEqualityFunction } from '../utils';
 
 class Stack<T> implements Iterable<T> {
   private list: LinkedList<T>;
